Extract test writer to new testlog package
